requirements.txt
streamlit==1.25.0
prophet==1.1.4
pandas==2.1.0
numpy==1.24.3
openai==0.27.8
python-dotenv==1.0.0
